refactor(autoapi): collapse specs package and relocate hook spec
