+++ pkgs/standards/auto_authn/auto_authn/orm/device_code.py
@@ -8,7 +8,7 @@ import uuid
 from autoapi.v3.orm.tables import Base
 from autoapi.v3.orm.mixins import Timestamped
 from autoapi.v3.specs import S, acol
-from autoapi.v3.specs.storage_spec import ForeignKeySpec
+from autoapi.v3.column.storage_spec import ForeignKeySpec
 from autoapi.v3.types import Boolean, Integer, Mapped, PgUUID, String, TZDateTime
 from autoapi.v3 import op_ctx
 from fastapi import HTTPException, status
