+++ pkgs/standards/autoapi/autoapi/v3/hook/__init__.py
@@ -3,7 +3,7 @@ from .decorators import hook_ctx
 from .types import PHASE, HookPhase, PHASES, Ctx, StepFn, HookPredicate
 from .shortcuts import hook, hook_spec
 from ._hook import Hook
-from ..hook_spec import HookSpec
+from .hook_spec import HookSpec
 
 __all__ = [
     "hook_ctx",
