+++ pkgs/standards/autoapi/autoapi/v3/specs.py
@@ -1,4 +1,4 @@
-# autoapi/v3/specs/__init__.py
+# autoapi/v3/specs.py
 """Compatibility layer that re-exports column specs.
 
 Importing from ``autoapi.v3.specs`` remains supported but the
@@ -7,8 +7,8 @@ implementation now lives under :mod:`autoapi.v3.column`.
 
 from __future__ import annotations
 
-from ..column import *  # noqa: F401,F403
-from ..column import __all__ as _all
+from .column import *  # noqa: F401,F403
+from .column import __all__ as _all
 
 __all__ = list(_all)
