--- pkgs/standards/autoapi/autoapi/v3/specs/column_spec.py
@@ -1,6 +1,0 @@
-# autoapi/v3/specs/column_spec.py
-"""Compatibility wrapper for :mod:`autoapi.v3.column.column_spec`."""
-
-from __future__ import annotations
-
-from ..column.column_spec import *  # noqa: F401,F403
