--- pkgs/standards/autoapi/autoapi/v3/specs/field_spec.py
@@ -1,6 +1,0 @@
-# autoapi/v3/specs/field_spec.py
-"""Compatibility wrapper for :mod:`autoapi.v3.column.field_spec`."""
-
-from __future__ import annotations
-
-from ..column.field_spec import *  # noqa: F401,F403
