--- pkgs/standards/autoapi/autoapi/v3/specs/infer.py
@@ -1,6 +1,0 @@
-# autoapi/v3/specs/infer.py
-"""Compatibility wrapper for :mod:`autoapi.v3.column.infer`."""
-
-from __future__ import annotations
-
-from ..column.infer import *  # noqa: F401,F403
