--- pkgs/standards/autoapi/autoapi/v3/specs/shortcuts.py
@@ -1,6 +1,0 @@
-# autoapi/v3/specs/shortcuts.py
-"""Compatibility wrapper for :mod:`autoapi.v3.column.shortcuts`."""
-
-from __future__ import annotations
-
-from ..column.shortcuts import *  # noqa: F401,F403
