--- pkgs/standards/autoapi/autoapi/v3/specs/storage_spec.py
@@ -1,6 +1,0 @@
-# autoapi/v3/specs/storage_spec.py
-"""Compatibility wrapper for :mod:`autoapi.v3.column.storage_spec`."""
-
-from __future__ import annotations
-
-from ..column.storage_spec import *  # noqa: F401,F403
