+++ pkgs/standards/autoapi/tests/unit/test_v3_storage_spec_attributes.py
@@ -2,7 +2,7 @@ from types import SimpleNamespace
 
 from autoapi.v3.runtime.atoms.storage import to_stored
 from autoapi.v3.specs import S, acol
-from autoapi.v3.specs.storage_spec import ForeignKeySpec, StorageTransform
+from autoapi.v3.column.storage_spec import ForeignKeySpec, StorageTransform
 from autoapi.v3.orm.tables import Base
 from sqlalchemy import Integer, String, text
 from sqlalchemy.orm import Mapped
