+++ pkgs/standards/peagen/peagen/orm/analysis_result.py
@@ -4,7 +4,7 @@ from autoapi.v3.orm.tables import Base
 from autoapi.v3.types import JSON, PgUUID, Text, Mapped, relationship
 from autoapi.v3.orm.mixins import GUIDPk, Timestamped, Ownable, TenantBound
 from autoapi.v3.specs import S, acol
-from autoapi.v3.specs.storage_spec import ForeignKeySpec
+from autoapi.v3.column.storage_spec import ForeignKeySpec
 from typing import TYPE_CHECKING
 
 from .users import User
