+++ pkgs/standards/peagen/peagen/orm/works.py
@@ -11,7 +11,7 @@ from autoapi.v3.types import (
 )
 from autoapi.v3.orm.mixins import GUIDPk, Timestamped, StatusColumn
 from autoapi.v3.specs import S, acol
-from autoapi.v3.specs.storage_spec import ForeignKeySpec
+from autoapi.v3.column.storage_spec import ForeignKeySpec
 from autoapi.v3 import hook_ctx
 from typing import TYPE_CHECKING
